更新：修复BUG，添加修改Bot信息的功能

+++ Telegram.Bot.Upgrader/Bot/BotManager.cs
@@ -14,37 +14,35 @@
 //  You should have received a copy of the GNU General Public License
 //  along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
-using System;
-using System.Collections.Generic;
-using System.IO;
 using System.IO.Compression;
-using System.Linq;
-using System.Reflection;
 using System.Text;
-using System.Threading.Tasks;
 using Telegram.Bot.Framework.Security;
 
 namespace Telegram.Bot.Upgrader.Bot
 {
     /// <summary>
-    /// 
+    /// Bot管理
     /// </summary>
     internal class BotManager : IBotManager, IDisposable
     {
+        // 存放Bot信息的文件夹
         private const string BOT_INFO_FOLDER = "BotInfos";
+        // Bot信息文件扩展名
         private const string BOT_INFO_PARAMNAME = ".botinfo";
-
+        // Bot执行文件目录
         private const string BOT_FILE_ROOT = "BotFileRoot";
 
         private static readonly Dictionary<string, BotInfoModel> BotInfo;
         static BotManager()
         {
             BotInfo = new Dictionary<string, BotInfoModel>();
+            _ = Directory.CreateDirectory(BOT_INFO_FOLDER);
+            // 从文件中加载Bot的信息
             foreach (string filePath in Directory.GetFiles(BOT_INFO_FOLDER, $"*{BOT_INFO_PARAMNAME}", SearchOption.TopDirectoryOnly))
             {
                 BotInfoModel botInfoModel = new BotInfoModel().Decode(File.ReadAllText(filePath));
                 botInfoModel.IsRunning = false;
-                BotInfo.Add(botInfoModel.BotID, botInfoModel);
+                BotInfo.Add(botInfoModel.BotID!, botInfoModel);
             }
         }
 
@@ -61,6 +59,11 @@ namespace Telegram.Bot.Upgrader.Bot
             Dispose();
         }
 
+        /// <summary>
+        /// 添加一个Bot
+        /// </summary>
+        /// <param name="botInfo">Bot信息</param>
+        /// <param name="botFile">Bot的可执行文件压缩包</param>
         public void AddBot(string botInfo, ZipArchive botFile)
         {
             // 生成ID
@@ -69,10 +72,11 @@ namespace Telegram.Bot.Upgrader.Bot
             DirectoryInfo directoryInfo = Directory.CreateDirectory(Path.Combine(BOT_FILE_ROOT, botID));
             // 讲压缩文件解压
             botFile.ExtractToDirectory(directoryInfo.FullName, true);
+            botFile.Dispose();
             // 找可执行文件
             FileInfo fileInfo = directoryInfo.GetFiles("*.exe", SearchOption.AllDirectories).Single();
 
-            BotInfoModel model = new BotInfoModel()
+            BotInfoModel model = new()
             {
                 BotID = botID,
                 BotInfo = botInfo,
@@ -83,11 +87,20 @@ namespace Telegram.Bot.Upgrader.Bot
             SaveBotInfo();
         }
 
-        public List<(string botID, string botInfo)> ListBotInfos()
+        /// <summary>
+        /// 列出所有的Bot
+        /// </summary>
+        /// <returns>Bot的信息列表</returns>
+        public List<BotInfoModel> ListBotInfos()
         {
-            return BotInfo.Select(x => (x.Value.BotID, x.Value.BotInfo)).ToList();
+            return BotInfo.Select(x => x.Value).ToList();
         }
 
+        /// <summary>
+        /// 移除Bot的信息，包括执行文件
+        /// </summary>
+        /// <param name="botID">Bot的ID</param>
+        /// <returns>是否执行成功</returns>
         public bool RemoveBot(string botID)
         {
             if (!BotInfo.ContainsKey(botID))
@@ -95,59 +108,86 @@ namespace Telegram.Bot.Upgrader.Bot
 
             BotInfoModel botInfo = BotInfo[botID];
 
-            StopBot(botInfo.BotID);
-            Directory.Delete(botInfo.BotFilePath, true);
+            _ = StopBot(botInfo.BotID!);
+            Directory.Delete(botInfo.BotFilePath!, true);
 
-            BotInfo.Remove(botID);
+            _ = BotInfo.Remove(botID);
             SaveBotInfo();
 
             return true;
         }
 
+        /// <summary>
+        /// 启动Bot
+        /// </summary>
+        /// <param name="botID">Bot的ID</param>
+        /// <returns>是否执行成功</returns>
         public bool StartBot(string botID)
         {
             if (!BotInfo.ContainsKey(botID))
                 return false;
 
             BotInfoModel botInfo = BotInfo[botID];
-            if (processManager.IsRunning(botInfo.BotID))
-            {
+            if (processManager.IsRunning(botInfo.BotID!))
                 return true;
-            }
-            botInfo.IsRunning = true;
-            return processManager.Start(botID, botInfo.EXEPath);
+            return File.Exists(botInfo.EXEPath) && (botInfo.IsRunning = processManager.Start(botID, botInfo.EXEPath));
         }
 
+        /// <summary>
+        /// 停止执行Bot
+        /// </summary>
+        /// <param name="botID">Bot的ID</param>
+        /// <returns>是否执行成功</returns>
         public bool StopBot(string botID)
         {
             if (!BotInfo.ContainsKey(botID))
                 return false;
 
             BotInfoModel botInfo = BotInfo[botID];
-            if (!processManager.IsRunning(botInfo.BotID))
-            {
+            if (!processManager.IsRunning(botInfo.BotID!))
                 return true;
-            }
             botInfo.IsRunning = false;
             return processManager.Stop(botID);
         }
 
+        /// <summary>
+        /// 升级Bot的执行文件
+        /// </summary>
+        /// <param name="botID">Bot的ID</param>
+        /// <param name="botFile">升级的文件</param>
+        /// <returns>是否执行成功</returns>
         public bool UpdateBot(string botID, ZipArchive botFile)
         {
             if (!BotInfo.ContainsKey(botID))
                 return false;
 
             BotInfoModel botInfo = BotInfo[botID];
-            
-            StopBot(botInfo.BotID);
-            Directory.Delete(botInfo.BotFilePath, true);
 
-            DirectoryInfo directoryInfo = Directory.CreateDirectory(botInfo.BotFilePath);
+            _ = StopBot(botInfo.BotID!);
+        RELOAD:
+            try
+            {
+                if (Directory.Exists(botInfo.BotFilePath))
+                    Directory.Delete(botInfo.BotFilePath, true);
+            }
+            catch (Exception)
+            {
+                Thread.Sleep(2000);
+                goto RELOAD;
+            }
+
+            DirectoryInfo directoryInfo = Directory.CreateDirectory(botInfo.BotFilePath!);
             botFile.ExtractToDirectory(directoryInfo.FullName, true);
 
             return true;
         }
 
+        /// <summary>
+        /// 更新Bot信息
+        /// </summary>
+        /// <param name="botID">Bot的ID</param>
+        /// <param name="botInfo">Bot的信息</param>
+        /// <returns>是否执行成功</returns>
         public bool UpdateBotInfo(string botID, string botInfo)
         {
             if (!BotInfo.ContainsKey(botID))
@@ -160,30 +200,60 @@ namespace Telegram.Bot.Upgrader.Bot
             return true;
         }
 
+        /// <summary>
+        /// 保存Bot信息
+        /// </summary>
         private static void SaveBotInfo()
         {
+            // 将信息更新到文件
             BotInfo.Select(x => x.Value).ToList().ForEach(model =>
             {
                 string filePath = Path.Combine(BOT_INFO_FOLDER, $"{model.BotID}{BOT_INFO_PARAMNAME}");
                 File.WriteAllText(filePath, model.Encrypt(), Encoding.UTF8);
             });
+            // 删除旧的文件
+            foreach (string filePath in Directory.GetFiles(BOT_INFO_FOLDER))
+                if (!BotInfo.ContainsKey(Path.GetFileNameWithoutExtension(filePath)))
+                    File.Delete(filePath);
         }
 
+        /// <summary>
+        /// 销毁
+        /// </summary>
         public void Dispose()
         {
             SaveBotInfo();
         }
     }
 
+    /// <summary>
+    /// Bot的相关信息
+    /// </summary>
     public class BotInfoModel : EncryptDecodeBase<BotInfoModel>
     {
-        public string BotID { get; set; }
-        public string BotInfo { get; set; }
+        /// <summary>
+        /// Bot的ID
+        /// </summary>
+        public string? BotID { get; set; }
 
-        public string BotFilePath { get; set; }
+        /// <summary>
+        /// Bot的信息
+        /// </summary>
+        public string? BotInfo { get; set; }
 
-        public string EXEPath { get; set; }
+        /// <summary>
+        /// Bot的文件夹路径
+        /// </summary>
+        public string? BotFilePath { get; set; }
 
+        /// <summary>
+        /// Bot执行文件路径
+        /// </summary>
+        public string? EXEPath { get; set; }
+
+        /// <summary>
+        /// Bot是否在运行中
+        /// </summary>
         public bool IsRunning { get; set; }
     }
 }
